+++ releasenotes.md
@@ -11,7 +11,7 @@
 * The player no longer continues to move across the map if they are moving when the automap is opened and the `am_followmode` CVAR is `off`.
 * A bug is fixed whereby the amount of cells the player had would always be brighter than usual in the widescreen HUD when their BFG-9000 was equipped.
 * The player’s view is no longer affected when walking over certain linedefs in a liquid sector and the `r_liquid_lowerview` CVAR is `on`.
-* The correct music now plays when playing *SIGIL* if `SIGIL_SHREDS.WAD` loads before `SIGIL.WAD` at startup.
+* Buckethead’s music is now heard when playing *SIGIL* and `SIGIL_SHREDS.WAD` is loaded before `SIGIL.WAD` at startup.
 * Obituaries displayed in the console when the player explodes a barrel using their BFG-9000 are now always correct.
 
 ![](https://github.com/bradharding/www.doomretro.com/raw/master/wiki/bigdivider.png)